requirements.txt
Django
django-widget-tweaks
djangorestframework
django-environ
pytz
psycopg2-binary
martor
gunicorn